Use CSV version of the database
CSV is easier to parse correctly than human-readable text.

This commit introduces a breaking change in the output format of the script.
The OUI for each organization is now printed without hyphens, as it's found in
the CSV database. Address normalization changes accordingly.

CSV parsing is implemented in Awk (dsv.awk preamble).

The argument to oui-lookup now permits but doesn't require separators.
The '.' separator is also allowed.